Refactoring to use shared-lib ipc
