+++ minion/ignite-worker/src/main/java/org/opennms/horizon/minion/identity/MinionIdentity.java
@@ -1,0 +1,13 @@
+package org.opennms.horizon.minion.identity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.opennms.horizon.shared.ipc.rpc.IpcIdentity;
+
+@Getter
+@AllArgsConstructor
+public class MinionIdentity implements IpcIdentity {
+
+    private final String id;
+    private final String location;
+}
