+++ minion/minion-features/pom.xml
@@ -45,9 +45,9 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>ipc-api</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.opennms.horizon.shared.ipc</groupId>
+            <artifactId>api</artifactId>
+            <version>${opennms-horizon.version}</version>
         </dependency>
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
@@ -76,20 +76,21 @@
             <version>${project.version}</version>
         </dependency>
 
+<!--        Duplicate?-->
+<!--        <dependency>-->
+<!--            <groupId>org.opennms.horizon.minion</groupId>-->
+<!--            <artifactId>ipc-api</artifactId>-->
+<!--            <version>${project.version}</version>-->
+<!--        </dependency>-->
         <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>ipc-api</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
+            <groupId>org.opennms.horizon.shared.ipc.grpc</groupId>
             <artifactId>client</artifactId>
-            <version>${project.version}</version>
+            <version>${opennms-horizon.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
+            <groupId>org.opennms.horizon.shared.ipc.grpc</groupId>
             <artifactId>contract</artifactId>
-            <version>${project.version}</version>
+            <version>${opennms-horizon.version}</version>
         </dependency>
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
@@ -173,8 +174,8 @@
                             </framework>
                             <features>
                                 <feature>ignite*</feature>
-                                <feature>poc-grpc*</feature>
-                                <feature>*-plugins</feature>
+<!--                                <feature>poc-grpc*</feature>-->
+<!--                                <feature>*-plugins</feature>-->
                             </features>
                         </configuration>
                     </execution>
