+++ minion/minion-grpc/workflow/module/pom.xml
@@ -18,9 +18,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>ipc-api</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.opennms.horizon.shared.ipc</groupId>
+            <artifactId>api</artifactId>
+            <version>${opennms-horizon.version}</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
