+++ minion/minion-grpc/workflow/workflow-subscriber/src/main/java/org/opennms/horizon/minion/ignite/grpc/subscriber/ResultsConsumer.java
@@ -11,8 +11,8 @@ import org.opennms.horizon.minion.ignite.grpc.workflow.contract.WorkflowResults.
 import org.opennms.horizon.minion.ignite.grpc.workflow.contract.WorkflowResults.WorkflowResult;
 import org.opennms.horizon.minion.ignite.model.workflows.Result;
 import org.opennms.horizon.minion.ignite.model.workflows.Results;
-import org.opennms.horizon.ipc.sink.api.MessageDispatcherFactory;
-import org.opennms.horizon.ipc.sink.api.SyncDispatcher;
+import org.opennms.horizon.shared.ipc.sink.api.MessageDispatcherFactory;
+import org.opennms.horizon.shared.ipc.sink.api.SyncDispatcher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.opennms.horizon.minion.ignite.grpc.workflow.WorkflowSinkModule;
