+++ minion/minion-grpc/workflow/workflow-subscriber/src/main/resources/OSGI-INF/blueprint/workflow-subscriber.xml
@@ -5,7 +5,7 @@
   xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
 
   <reference id="identity" interface="org.opennms.horizon.core.identity.Identity" />
-  <reference id="messageDispatcherFactory" interface="org.opennms.horizon.ipc.sink.api.MessageDispatcherFactory" />
+  <reference id="messageDispatcherFactory" interface="org.opennms.horizon.shared.ipc.sink.api.MessageDispatcherFactory" />
 
   <bean id="subscriber" class="org.opennms.horizon.minion.ignite.grpc.subscriber.LoggingWorkflowSubscriber" />
