+++ minion/plugins/registration/src/main/java/org/opennms/horizon/minion/registration/PluginConfigSinkModule.java
@@ -1,10 +1,10 @@
 package org.opennms.horizon.minion.registration;
 
-import org.opennms.horizon.ipc.sink.aggregation.IdentityAggregationPolicy;
-import org.opennms.horizon.ipc.sink.api.AggregationPolicy;
-import org.opennms.horizon.ipc.sink.api.AsyncPolicy;
-import org.opennms.horizon.ipc.sink.api.SinkModule;
 import org.opennms.horizon.minion.registration.proto.PluginConfigMessage;
+import org.opennms.horizon.shared.ipc.sink.aggregation.IdentityAggregationPolicy;
+import org.opennms.horizon.shared.ipc.sink.api.AggregationPolicy;
+import org.opennms.horizon.shared.ipc.sink.api.AsyncPolicy;
+import org.opennms.horizon.shared.ipc.sink.api.SinkModule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
