+++ minion/plugins/registration/src/main/resources/OSGI-INF/blueprint/registration.xml
@@ -15,7 +15,7 @@
     <camel:template id="registrationProducer" defaultEndpoint="{{routing.uri}}"/>
   </camel:camelContext>
 
-  <reference id="messageDispatcherFactory" interface="org.opennms.horizon.ipc.sink.api.MessageDispatcherFactory" />
+  <reference id="messageDispatcherFactory" interface="org.opennms.horizon.shared.ipc.sink.api.MessageDispatcherFactory" />
 
   <bean id="registrationService" class = "org.opennms.horizon.minion.registration.RegistrationServiceImpl">
     <argument ref="registrationProducer"/>
