+++ parent-pom/pom.xml
@@ -15,7 +15,7 @@
     </description>
 
     <properties>
-        <dropwizard.metrics.version>3.1.2</dropwizard.metrics.version>
+        <dropwizard.metrics.version>4.2.6</dropwizard.metrics.version>
         <activemq.version>5.17.1</activemq.version>
         <aries.jax.rs.whiteboard.version>2.0.1</aries.jax.rs.whiteboard.version>
         <aries.jpa.api.version>2.7.3</aries.jpa.api.version>
